requirements.txt
aiogram==3.13.1
openai>=1.30.0
Pillow>=10.0.0
python-dotenv>=1.0.0
